% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_rank_plot.R
\name{bc_rank_plot}
\alias{bc_rank_plot}
\title{Barcode Rank Plot}
\usage{
bc_rank_plot(lnc_gg, save)
}
\arguments{
\item{lnc_gg}{Datafrome that contains log (number of counts), rank and the whether the cell is selected for plotting}

\item{save}{Output directory for the plot}
}
\description{
Barcode Rank Plot
}
