% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.anno.at.R
\name{cor.anno.at}
\alias{cor.anno.at}
\title{Correlation-based annotation for Arabidopsis Root}
\usage{
cor.anno.at(
  seu,
  dir_to_bulk,
  unwanted.genes,
  clustering_alg,
  res,
  mt.pattern,
  cp.pattern
)
}
\arguments{
\item{seu}{Seurat object}

\item{dir_to_bulk}{Directory to reference expression profile}

\item{unwanted.genes}{Unwanted genes to exclude from analysis (character vector)}

\item{clustering_alg}{Algorithm for clustering (1 = original Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm; 4 = Leiden algorithm, which requires the leidenalg python; numeric)}

\item{res}{Resolution used for Leiden clustering (numeric)}

\item{mt.pattern}{Pattern of mitochondrial gene names/ids (character) ex. "ATMG" or list of mitochondrial genes (character vector)}

\item{cp.pattern}{Pattern of chloroplast gene names/ids (character) ex. "ATCG" or list of chloroplast genes (character vector)}
}
\description{
Correlation-based annotation for Arabidopsis Root
}
