% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.doub.rate.R
\name{est.doub.rate}
\alias{est.doub.rate}
\title{Estimated doublet rate according to 10X genomics}
\usage{
est.doub.rate(ncell)
}
\arguments{
\item{ncell}{number of high quality cells remain after quality filtering}
}
\description{
Estimated doublet rate according to 10X genomics
}
