% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{feature_plot}
\alias{feature_plot}
\title{Feature plot}
\usage{
feature_plot(seu, res, doublet.rate, save)
}
\arguments{
\item{seu}{Seurat object}

\item{res}{Resolution used for clustering (numeric)}

\item{doublet.rate}{Doublet rate estimated (numeric)}

\item{save}{Output directory for the feature plot}
}
\description{
Feature plot
}
