% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot_at.R
\name{feature_plot_at}
\alias{feature_plot_at}
\title{Feature plot for Arabidopsis Root}
\usage{
feature_plot_at(seu, res, doublet.rate, dir_to_color_scheme, save)
}
\arguments{
\item{seu}{Seurat object}

\item{res}{Resolution used for clustering (numeric)}

\item{doublet.rate}{Doublet rate estimated (numeric)}

\item{dir_to_color_scheme}{Directory to color scheme}

\item{save}{Output directory for the feature plot}
}
\description{
Feature plot for Arabidopsis Root
}
