% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_hist_plot.R
\name{Gene_hist_plot}
\alias{Gene_hist_plot}
\title{Gene histogram}
\usage{
Gene_hist_plot(lng_gg, save)
}
\arguments{
\item{save}{Output directory for the plot}

\item{lnc_gg}{Datafrome that contains log (number of genes), rank and the whether the cell is selected for plotting}
}
\description{
Gene histogram
}
