% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.rownames.R
\name{merge.rownames}
\alias{merge.rownames}
\title{Merge by row names}
\usage{
\method{merge}{rownames}(x, y)
}
\arguments{
\item{x}{The first matrix}

\item{y}{The second matrix}
}
\description{
Merge by row names
}
