% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_HTML_no_analysis.R
\name{print_HTML_no_analysis}
\alias{print_HTML_no_analysis}
\title{Print HTML for users who do not want analysis by Seurat}
\usage{
print_HTML_no_analysis(
  parameters,
  cell_stats,
  seq_stats,
  sample_stats,
  dir,
  sample.name
)
}
\arguments{
\item{parameters}{Parameters used in copilot_prepros function}

\item{cell_stats}{Cell statistics}

\item{seq_stats}{Sequencing statistics}

\item{sample_stats}{Sample statistics}

\item{dir}{Directory to the plots}

\item{sample.name}{Sample name}
}
\description{
Print HTML for users who do not want analysis by Seurat
}
