% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMI_hist_plot.R
\name{UMI_hist_plot}
\alias{UMI_hist_plot}
\title{UMI histogram}
\usage{
UMI_hist_plot(lnc_gg, save, legend.position)
}
\arguments{
\item{lnc_gg}{Datafrome that contains log (number of counts), rank and the whether the cell is selected for plotting}

\item{save}{Output directory for the plot}

\item{lengend.position}{x y position of the legend, e.g. c(0.8, 0.8)}
}
\description{
UMI histogram
}
